backend/requirements-minimal.txt
# Core FastAPI Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23

# Environment & Configuration
python-dotenv==1.0.0

# Web & HTTP
requests==2.31.0
websockets==12.0

# Authentication
bcrypt==4.0.1
passlib==1.7.4
python-jose[cryptography]==3.3.0

# File Processing (minimal)
Pillow==10.1.0
